+++ src/TrackExpressionVars.h
@@ -12,6 +12,7 @@
 #include <vector>
 #include <string>
 #include <string.h>
+#include <memory>
 
 #ifndef R_NO_REMAP
 #  define R_NO_REMAP
@@ -25,6 +26,7 @@
 #include "GInterval.h"
 #include "GInterval2D.h"
 #include "TrackExpressionIteratorBase.h"
+#include "PWMScorer.h"
 #include "rdbinterval.h"
 #include "rdbutils.h"
 
@@ -85,7 +87,7 @@ public:
     typedef vector<Track_n_imdf> Track_n_imdfs;
 
     struct Track_var {
-        enum Val_func { REG, REG_MIN, REG_MAX, REG_NEAREST, STDDEV, SUM, QUANTILE, PV, PV_MIN, PV_MAX, WEIGHTED_SUM, OCCUPIED_AREA, NUM_FUNCS };
+        enum Val_func { REG, REG_MIN, REG_MAX, REG_NEAREST, STDDEV, SUM, QUANTILE, PV, PV_MIN, PV_MAX, WEIGHTED_SUM, OCCUPIED_AREA, PWM, PWM_MAX, PWM_MAX_POS, NUM_FUNCS };
 
         static const char *FUNC_NAMES[NUM_FUNCS];
 
@@ -97,6 +99,8 @@ public:
         bool                requires_pv;
         Binned_pv           pv_binned;
         Track_n_imdf       *track_n_imdf;
+        std::unique_ptr<PWMScorer> pwm_scorer;
+        char strand;
     };
 
     typedef vector<Track_var> Track_vars;
@@ -135,12 +139,14 @@ public:
 	void define_r_vars(unsigned size);
     const Track_var *var(const char *var_name) const;
 
+    bool is_pwm_variable(unsigned ivar) const;
+
 	void set_vars(const GInterval &interval, unsigned idx);
 	void set_vars(const GInterval2D &interval, const DiagonalBand &band, unsigned idx);
 
 private:
-	rdb::IntervUtils       &m_iu;
-	string                  m_groot;
+    rdb::IntervUtils &m_iu;
+    string                  m_groot;
 	Track_vars              m_track_vars;
 	Interv_vars             m_interv_vars;
 	Track_n_imdfs           m_track_n_imdfs;
@@ -167,6 +173,19 @@ private:
 	void set_vars(unsigned idx);
 
 	bool is_var(const string &str, uint64_t start, uint64_t end) const { return (!start || !rdb::is_R_var_char(str[start - 1])) && (end == str.size() || !rdb::is_R_var_char(str[end])); }
+
+    static int findListElementIndex(SEXP list, const char* name) {
+        SEXP names = Rf_getAttrib(list, R_NamesSymbol);
+        if (names == R_NilValue)
+            rdb::verror("List must have named elements");
+            
+        int len = Rf_length(list);
+        for (int i = 0; i < len; i++) {
+            if (strcmp(CHAR(STRING_ELT(names, i)), name) == 0)
+                return i;
+        }
+        return -1;  // Element not found
+    }
 };
 
 
@@ -215,6 +234,12 @@ inline const TrackExpressionVars::Track_var *TrackExpressionVars::var(const char
             return &*ivar;
     }
     return NULL;
+}
+
+inline bool TrackExpressionVars::is_pwm_variable(unsigned ivar) const {
+    return m_track_vars[ivar].val_func == Track_var::PWM ||
+           m_track_vars[ivar].val_func == Track_var::PWM_MAX ||
+           m_track_vars[ivar].val_func == Track_var::PWM_MAX_POS;
 }
 
 #endif /* TRACKEXPRESSIONVARS_H_ */